Refactor acquisition team sub-form to be shared
